Recover handlers in `LocalRPC`
